Downloading resources only once

+++ ground-station/build.sh
@@ -14,28 +14,36 @@ else
     ARDUINO_DIR="arduino-$ARDUINO_IDE_VERSION"
 fi
 
-ARDUINO_FILE=arduino-${ARDUINO_IDE_VERSION}-$PLATFORM.$FILE_EXT
-wget -O $ARDUINO_FILE http://arduino.cc/download.php?f=/$ARDUINO_FILE
-
-if [ $FILE_EXT == 'zip' ]
+if [ ! -e "$ARDUINO_DIR" ]
 then
-    unzip $ARDUINO_FILE
-else
-    tar xf $ARDUINO_FILE
-fi
+    ARDUINO_FILE=arduino-${ARDUINO_IDE_VERSION}-$PLATFORM.$FILE_EXT
+    wget -O $ARDUINO_FILE http://arduino.cc/download.php?f=/$ARDUINO_FILE
+
+    if [ $FILE_EXT == 'zip' ]
+    then
+        unzip $ARDUINO_FILE
+    else
+        tar xf $ARDUINO_FILE
+    fi
 
 rm -f $ARDUINO_FILE
-echo "ARDUINO_DIR = $ARDUINO_DIR" >> Makefile
-
-FILE=$MAKEFILE_VERSION.tar.gz
-
-wget https://github.com/sudar/Arduino-Makefile/archive/$FILE
-tar xzf $FILE
-rm -f $FILE
+fi
 
 ARDMK_DIR=Arduino-Makefile-$MAKEFILE_VERSION
 
-echo "CPPFLAGS = -DDEBUG=$DEBUG" >> Makefile
-echo "include $ARDMK_DIR/Arduino.mk" >> Makefile
+if [ ! -e $ARDMK_DIR ]
+then
+    FILE=$MAKEFILE_VERSION.tar.gz
+    wget https://github.com/sudar/Arduino-Makefile/archive/$FILE
+    tar xzf $FILE
+    rm -f $FILE
+fi
+
+if [ ! -e "Makefile" ]
+then
+    echo "ARDUINO_DIR = $ARDUINO_DIR" >> Makefile
+    echo "CPPFLAGS = -DDEBUG=$DEBUG" >> Makefile
+    echo "include $ARDMK_DIR/Arduino.mk" >> Makefile
+fi
 
 make DEBUG=$DEBUG
